requirements.txt
# goFaster v1
asgiref==3.7.2
Django==4.1.7
django-cors-headers==4.2.0 
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.